[FIX] l10n_ve_igt, l10n_ve_accountat:  Correcciones
- En contabilidad se elimina conversion numero 6 de las lineas de asientos en base al monto de la sretenciones, este era un enfoque q se heredo de V17 el cuel era incorrecto, dichas lineas van de igual forma convertidas como cualquier otro pago.

- Se elimina _compute_amount_currency  en contabilidad , dicha funcion en contabilidad sobreescribia la nativa y pasaba por alto la porcion real al reincorporar el error de redondeo por conversion directa.

- Se Agregando condiciones para ejecutar las distribuciones de la porcion real solamente cuando se este en estado de borrador ya que cualquier otro campo podria disparar estos calculos erroneamente

- Se mueve el ajuste de  _inverse_amount_currency ubicado en l10nve_igtf hacia l10n_ve_accountant ya que dicho ajuste deberia mantenerse en esa parte en especifico.

- Se corrijen campos en l10n_ve_igtf, se eleiminan logicas obsoletas de recalculo del  _compute_tax_totals el cual dependia del bi_igtf

References:
- Ticket: https://binaural.odoo.com/odoo/helpdesk/action-389/13513
- Tarea:
- PR:
- Otros:
