+++ l10n_ve_igtf/models/account_move_line.py
@@ -1,5 +1,4 @@
 from odoo import api, fields, models, _
-from odoo.tools import float_round, float_is_zero
 from odoo.exceptions import UserError
 
 import logging
@@ -13,52 +12,7 @@ class AccountMoveLine(models.Model):
         string="Payment Advance"
     )
 
-    @api.onchange('amount_currency', 'currency_id')
-    def _inverse_amount_currency(self):
-        """
-        Updates the 'balance' (company currency amount) whenever the 'amount_currency' 
-        or 'currency_id' changes, ensuring a symmetric rounding.
-
-        This method addresses the common floating-point discrepancy where a balance 
-        converted to foreign currency and then back to company currency results in 
-        a small difference (e.g., 0.01). 
-
-        The logic performs a "Symmetry Test":
-        1. It calculates the initial balance using the current exchange rate.
-        2. It simulates a back-conversion to the foreign currency.
-        3. If the back-conversion doesn't match the original 'amount_currency' due to 
-        rounding noise, it applies a micro-adjustment to the 'balance' in the 
-        company currency (VES) to force a perfect match.
-
-        :return: None
-        """
-        for line in self:
-            if line.currency_id == line.company_id.currency_id and line.balance != line.amount_currency:
-                line.balance = line.amount_currency
-                
-            elif (
-                line.currency_id != line.company_id.currency_id
-                and not line.move_id.is_invoice(True)
-                and not self.env.is_protected(self._fields['balance'], line)
-            ):
-                rate = line.currency_rate
-                if not rate:
-                    continue
-                    
-                raw_balance = line.amount_currency / rate
-                
-                rounded_balance = line.company_id.currency_id.round(raw_balance)
-                
-                back_to_foreign = rounded_balance * rate
-                diff_foreign = line.amount_currency - back_to_foreign
-                
-                if not float_is_zero(diff_foreign, precision_rounding=line.currency_id.rounding):
-                    adjustment = float_round(diff_foreign / rate, precision_rounding=line.company_id.currency_id.rounding)
-                    line.balance = rounded_balance + adjustment
-                else:
-                    line.balance = rounded_balance
-
-
+    
     def action_register_payment(self):
         """ 
         # 1. Validate Unique Partner
